refactor&feat: moved files to parser/antlr package and added better error handling
