+++ pom.xml
@@ -163,7 +163,7 @@
                     <outputDirectory>src/main/java</outputDirectory>
                     <arguments>
                         <argument>-visitor</argument>
-                        <argument>calculator/parser/LabeledExpr.g4</argument>
+                        <argument>calculator/parser/antlr/LabeledExpr.g4</argument>
                         <!-- <argument>-listener</argument> -->
                     </arguments>
                 </configuration>
