+++ src/main/java/calculator/Main.java
@@ -1,8 +1,6 @@
 package calculator;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import calculator.parser.CalculatorParser;
 
 /**
  * A very simple calculator in Java
@@ -27,41 +25,12 @@ public class Main {
 
 	try{
 
-		e = new MyNumber(8);
-		c.print(e);
-		c.eval(e);
-
-	    List<Expression> params = new ArrayList<>();
-	    Collections.addAll(params, new MyNumber(3), new MyNumber(4), new MyNumber(5));
-	    e = new Plus(params,Notation.PREFIX);
-		c.printExpressionDetails(e);
-		c.eval(e);
-	
-		List<Expression> params2 = new ArrayList<>();
-		Collections.addAll(params2, new MyNumber(5), new MyNumber(3));
-		e = new Minus(params2, Notation.INFIX);
-		c.print(e);
-		c.eval(e);
-
-		List<Expression> params3 = new ArrayList<>();
-		Collections.addAll(params3, new Plus(params), new Minus(params2));
-		e = new Times(params3);
-		c.printExpressionDetails(e);
-		c.eval(e);
-
-		List<Expression> params4 = new ArrayList<>();
-		Collections.addAll(params4, new Plus(params), new Minus(params2), new MyNumber(5));
-		e = new Divides(params4,Notation.POSTFIX);
-		c.print(e);
-		c.eval(e);
-
-
-		Expression exp = CalculatorParser.parseString("(((4,5,6)+,(7,(5,2,7)/)+,9) * 2) /");
+		Expression exp = CalculatorParser.parseString("(4 + )");
 		c.printExpressionDetails(exp);
 	}
 
-	catch(IllegalConstruction exception) {
-		System.out.println("cannot create operations without parameters");
+	catch(Exception exception) {
+		System.out.println("cannot create operations without parameters: " + exception.getMessage());
 		}
  	}
