+++ src/main/java/calculator/parser/ThrowingErrorListener.java
@@ -1,0 +1,23 @@
+package calculator.parser;
+
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.Parser;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
+import org.antlr.v4.runtime.atn.ATNConfigSet;
+import org.antlr.v4.runtime.dfa.DFA;
+import org.antlr.v4.runtime.misc.ParseCancellationException;
+
+
+/**
+ * Simple class meant to catch problems detected by the parser and throw an error encapsulating the important information.
+ */
+public class ThrowingErrorListener extends BaseErrorListener
+{
+    @Override
+    public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line, int charPositionInLine, String msg, RecognitionException e)
+            throws ParseCancellationException
+    {
+        throw new ParserError(line, charPositionInLine, msg);
+    }
+}
