+++ src/main/java/calculator/parser/antlr/LabeledExprParser.java
@@ -1,5 +1,5 @@
-// Generated from calculator/parser/LabeledExpr.g4 by ANTLR 4.9.1
-package calculator.parser;
+// Generated from calculator/parser/antlr/LabeledExpr.g4 by ANTLR 4.9.1
+package calculator.parser.antlr;
 import org.antlr.v4.runtime.atn.*;
 import org.antlr.v4.runtime.dfa.DFA;
 import org.antlr.v4.runtime.*;
